Move all configuration related modules to `config` package
