+++ src/statue/cli/contexts.py
@@ -4,7 +4,7 @@ import click
 
 from statue.cli.cli import statue_cli
 from statue.cli.styled_strings import bullet_style, name_style
-from statue.configuration import Configuration
+from statue.config.configuration import Configuration
 from statue.exceptions import UnknownContext
